[fx/package] make GraphModules packageable
FX serializes things by serializing Python code as a string and exec'ing
it on load. This accomplishes one goal (we don't have to pickle the
graph object directly) but breaks the pickle abstraction in ways that
are not composable with `torch.package`.

In particular:
1. `forward` is serialized by saving Python code. On load, it's installed
by  `exec`ing that code. This `exec` call needs to have the right
importer installed, otherwise it will not import modules from the
`torch.package` but instead import from the Python environment.
2. Any types/functions used are emitted as `import` statement in the
generated Python code. These are effectively dynamic dependencies of the
`GraphModule` being saved, and need to be registered as such so that the
`PackageImporter` will package them.

This PR introduces some new concepts to torch.package to fix these
problem.

- `ImporterProxy`: When unpickled by a PackageImporter, this object will
trun into the PackageImporter instance itself. This lets `__reduce__`
function authors access the current PackageImporter instance in its
reduction function. This is accomplished by `persistent_id` hackery.
This solves problem #1 above.
- `__torch_package_persist__`: `persistent_id` in PackageExporter will
check for this method and run it if it exists, passing in the
PacakgeExporter itself as an argument. We use this basically as a hook
function that runs whenenver a GraphModule instance is pickled.  This
hook registers the necessary imports, solving problem #2 above.

There are some follow ups, in particular, you can't really return a
non-None value from `__torch_package_persist__` because we don't
currently have a way to hook into `persistent_load` on the other side,
so you're limited to running side-effectful hooks.

ghstack-source-id: e5a196906b92ee4fc92e351b625cba9168411687
Pull Request resolved: https://github.com/pytorch/pytorch/pull/51489